move melody, drums, section to composition/ with Section validation
Section.__init__ now validates chord progression strings at
construction time via parse_roman_numeral(). Raises ValueError
on malformed input instead of failing deep in compilation.
