+++ midigen/composition/__init__.py
@@ -3,3 +3,6 @@
 from midigen.composition.chord import Chord, CHORD_TYPES
 from midigen.composition.progression import ChordProgression
 from midigen.composition.arpeggio import Arpeggio, ArpeggioPattern
+from midigen.composition.melody import Melody
+from midigen.composition.drums import DrumKit, Drum, GM1_DRUM_MAP
+from midigen.composition.section import Section
