+++ midigen/composition/melody.py
