--- midigen/section.py
@@ -1,5 +1,0 @@
-class Section:
-    def __init__(self, name: str, length: int, chord_progression: str):
-        self.name = name
-        self.length = length
-        self.chord_progression = chord_progression
